# filter: prevent escape of QDecode to the heap