--- a/wgengine/filter/filter.go
+++ b/wgengine/filter/filter.go
@@ -138,16 +138,16 @@ var acceptBucket = rate.NewLimiter(rate.Every(10*time.Second), 3)
 var dropBucket = rate.NewLimiter(rate.Every(5*time.Second), 10)
 
 func (f *Filter) logRateLimit(runflags RunFlags, b []byte, q *packet.QDecode, r Response, why string) {
+	var qs string
+	if q == nil {
+		qs = fmt.Sprintf("(%d bytes)", len(b))
+	} else {
+		qs = q.String()
+	}
 	if r == Drop && (runflags&LogDrops) != 0 && dropBucket.Allow() {
-		var qs string
-		if q == nil {
-			qs = fmt.Sprintf("(%d bytes)", len(b))
-		} else {
-			qs = q.String()
-		}
 		f.logf("Drop: %v %v %s\n%s", qs, len(b), why, maybeHexdump(runflags&HexdumpDrops, b))
 	} else if r == Accept && (runflags&LogAccepts) != 0 && acceptBucket.Allow() {
-		f.logf("Accept: %v %v %s\n%s", q, len(b), why, maybeHexdump(runflags&HexdumpAccepts, b))
+		f.logf("Accept: %v %v %s\n%s", qs, len(b), why, maybeHexdump(runflags&HexdumpAccepts, b))
 	}
 }