--- a/wgengine/filter/filter.go
+++ b/wgengine/filter/filter.go
@@ -138,24 +138,26 @@ var acceptBucket = rate.NewLimiter(rate.Every(10*time.Second), 3)
 var dropBucket = rate.NewLimiter(rate.Every(5*time.Second), 10)
 
 func (f *Filter) logRateLimit(runflags RunFlags, b []byte, q *packet.QDecode, r Response, why string) {
-	// Note: it is crucial that q.String() be called only after {accept,drop}Bucket.Allow() passes,
-	// since it causes an allocation.
+	var verdict string
+
 	if r == Drop && (runflags&LogDrops) != 0 && dropBucket.Allow() {
-		var qs string
-		if q == nil {
-			qs = fmt.Sprintf("(%d bytes)", len(b))
-		} else {
-			qs = q.String()
-		}
-		f.logf("Drop: %v %v %s\n%s", qs, len(b), why, maybeHexdump(runflags&HexdumpDrops, b))
+		verdict = "Drop"
+		runflags &= HexdumpDrops
 	} else if r == Accept && (runflags&LogAccepts) != 0 && acceptBucket.Allow() {
+		verdict = "Accept"
+		runflags &= HexdumpAccepts
+	}
+
+	// Note: it is crucial that q.String() be called only if {accept,drop}Bucket.Allow() passes,
+	// since it causes an allocation.
+	if verdict != "" {
 		var qs string
 		if q == nil {
 			qs = fmt.Sprintf("(%d bytes)", len(b))
 		} else {
 			qs = q.String()
 		}
-		f.logf("Accept: %v %v %s\n%s", qs, len(b), why, maybeHexdump(runflags&HexdumpAccepts, b))
+		f.logf("%s: %s %d %s\n%s", verdict, qs, len(b), why, maybeHexdump(runflags, b))
 	}
 }