Right now, `q` from `tstun` escapes to the heap, as reported by
```sh
go install --gcflags=tailscale.com/wgengine/...=-m ./wgengine/tstun
```
Namely, we have
```
wgengine/tstun/tun.go:179:6: moved to heap: q
```
in
```go
179	var q packet.QDecode
180	if filt.RunOut(buf, &q, t.filterFlags) == filter.Accept {
```
and
```
wgengine/tstun/tun.go:221:6: moved to heap: q
```
in
```go
221	var q packet.QDecode
222	if filt.RunIn(buf, &q, t.filterFlags) == filter.Accept {
```
In turns out that this is because `q` is directly passed to `logf` in one case, and `logf` is an interface, so escape is inevitable. This slight reordering fixes that.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tailscale/tailscale/417)
<!-- Reviewable:end -->