@bradfitz 2020-05-27
Could you add a test to make sure this doesn't regress in the future?

And perhaps also a `BenchmarkFoo(b *testing.B)` func?