@bradfitz 2020-05-27
> Do we have any way to add a test to prevent a regression? I don't know what our options are for escape analysis in automation.

testing.AllocsPerRun should catch it.