@dshynkev 2020-05-28
Caught another allocation. In total, we have:
```
======
BEFORE
======
goos: linux
goarch: amd64
pkg: tailscale.com/wgengine/filter
BenchmarkFilter/tcp-4         	14656930	       73.2 ns/op
BenchmarkFilter/udp-4         	11694705	      100 ns/op
PASS
ok  	tailscale.com/wgengine/filter	2.437s

=====
AFTER
=====
goos: linux
goarch: amd64
pkg: tailscale.com/wgengine/filter
BenchmarkFilter/tcp-4         	38456259	       30.3 ns/op
BenchmarkFilter/udp-4         	21849141	       54.1 ns/op
PASS
ok  	tailscale.com/wgengine/filter	2.437s
```
Very importantly, what I initially proposed what much worse, since it called `q.String()` every time, regardless of rate-limiting, so adding a benchmark really mattered:
```
goos: linux
goarch: amd64
pkg: tailscale.com/wgengine/filter
BenchmarkFilter/tcp-4         	3633560	      329 ns/op
BenchmarkFilter/udp-4         	3377547	      351 ns/op
PASS
ok  	tailscale.com/wgengine/filter	3.076s
```