@bradfitz 2020-05-28
Btw, for comparing benchmark results between BEFORE & AFTER, the common Go tool used is:

https://godoc.org/golang.org/x/perf/cmd/benchstat